requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
websockets>=12.0
httpx>=0.25.0

# Voice processing
openai-whisper>=20231117
piper-tts>=1.2.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
librosa>=0.10.0

# AI and ML
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
scikit-learn>=1.3.0

# Home automation
homeassistant-api>=3.0.0
python-telegram-bot>=20.0.0
paho-mqtt>=1.6.0

# Data processing
pandas>=2.0.0
pillow>=10.0.0
requests>=2.31.0
jsonschema>=4.19.0
opencv-python>=4.8.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0
psutil>=5.9.0